fix: folding html element will include attributes and close element
For html-mode and html-ts-mode, without this change, treesit-fold folds from
the END (closing angle bracket) of the opening element to the beginning (opening
angle bracket) of the close element. That means attributes on the element are
not included in the fold. Modern HTML can include lots of attributes.

Folding from the END of the opening element name to the END (closing
angle bracket) of the closing element, will include attributes in the fold.

This change does that.